change sequence of colors from randrom to yellow/blue/red
